So many changes, things should work now.
I made it so that you are passing around double pointers, which will
actually pass your matricies around instead of passing a single
(illegal) value from your matrix to out of it.  I also changed your code
so that you are using the value out of your B and D matricies instead of
copying them into a local value.  Also, you were using inheritance for
some reason, and I don't think it is what you inteded to be doing.