trouble pulling in K matrices
for this part of the code i need to pull in the K matrix for each
element, then pull each value from the K matrix in turn and place it in
the correct spot in the global matrix, which will ultimately include all
the values from all the K matrices.

I get the following error at the part of the code where I am trying to
get the K matrix:
error C2109: subscript requires array or pointer type

Based on how the k matrices are called, i'm not sure where the problem
is.