Added a new feature (with its own problems)
this push has the same problem pulling the K matrix in the
global_matrix.cpp that the last one did, but i added a function to
create a load vector.  The load vector builds, but when I run the
program and try to print the output, I get an error.